Add tomorrow font and refactor fonts for specific elements
