+++ versaepoch/src/app/layout.tsx
@@ -1,21 +1,12 @@
 import type { Metadata } from 'next';
-import { Geist, Geist_Mono } from 'next/font/google';
 import { Header } from '@/components/Header';
 import { Footer } from '@/components/Footer';
 import { ScrollToTop } from '@/components/ScrollToTop';
 import VercelAnalyticsClient from '@/components/AnalyticsClient';
 import { SpeedInsights } from '@vercel/speed-insights/next';
+import { tomorrow, geistMono, geistSans } from './fonts';
 import './globals.css';
 
-const geistSans = Geist({
-  variable: '--font-geist-sans',
-  subsets: ['latin'],
-});
-
-const geistMono = Geist_Mono({
-  variable: '--font-geist-mono',
-  subsets: ['latin'],
-});
 
 export const metadata: Metadata = {
   title: 'VersaEpoch - Explore Evolution of ChatGPT, Claude & Gemini',
@@ -58,7 +49,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body
-        className={`${geistSans.variable} ${geistMono.variable} rootContainer antialiased`}>
+        className={`${geistSans.variable} ${geistMono.variable} ${tomorrow.variable} rootContainer antialiased`}>
         <ScrollToTop />
         <Header />
         {children}
