+++ versaepoch/src/styles/ui/tableHeaderDropdown.module.scss
@@ -31,6 +31,7 @@
 .headerTitle {
   flex: 1;
   text-align: left;
+  font-family: var(--font-geist-mono);
 }
 
 .sortIcon {
